Modularization done
