+++ middlewares/authMiddleware.js
@@ -1,0 +1,8 @@
+// File: /middlewares/authMiddleware.js
+
+export const ensureAuthenticated = (req, res, next) => {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect('/login');
+};
