+++ services/dbClient.js
@@ -1,0 +1,23 @@
+// File: /services/dbClient.js
+
+import pg from 'pg';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const db = new pg.Client({
+    user: process.env.POSTGRESQL_ADDON_USER,
+    host: process.env.POSTGRESQL_ADDON_HOST,
+    database: process.env.POSTGRESQL_ADDON_DB,
+    password: process.env.POSTGRESQL_ADDON_PASSWORD,
+    port: process.env.POSTGRESQL_ADDON_PORT,
+    // user: process.env.DB_USER,
+    // host: process.env.DB_HOST,
+    // database: process.env.DB_NAME,
+    // password: process.env.DB_PASSWORD,
+    // port: process.env.DB_PORT,
+});
+
+db.connect();
+
+export default db;
