+++ views/draft_list.ejs
@@ -279,7 +279,6 @@
               <a class="blog-header-logo text-body-emphasis text-decoration-none" href="/">Blog.ai</a>
             </div>
             <div class="col-3 d-flex justify-content-end align-items-center">
-                <span class="link-secondary">Subscribed</span>
                 <a class="btn btn-sm btn-outline-secondary ms-4" href="/logout">Log out</a>
             <!-- Theme Toggle Button -->
         <div class="dropdown ms-4 bd-mode-toggle">
