+++ views/header.ejs
@@ -32,11 +32,6 @@
       </div>
       <div class="col-3 d-flex justify-content-end align-items-center">
         <% if (user) { %>
-          <span class="link-secondary">Subscribed</span>
-        <% } else { %>
-          <a class="link-secondary" href="/subscribe">Subscribe</a>
-        <% } %>
-        <% if (user) { %>
           <a class="btn btn-outline-secondary py-2 ms-3" href="/profile">
             <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" fill="currentColor" class="bi bi-person-fill" viewBox="0 0 16 16">
               <path d="M3 14s-1 0-1-1 1-4 6-4 6 3 6 4-1 1-1 1zm5-6a3 3 0 1 0 0-6 3 3 0 0 0 0 6"/>
