+++ views/profile.ejs
@@ -279,8 +279,6 @@
           </div>
           <div class="col-3 d-flex justify-content-end align-items-center">
             
-              <span class="link-secondary">Subscribed</span>
-            
               <a class="btn btn-sm btn-outline-secondary ms-4" href="/logout">Log out</a>
             
           <!-- Theme Toggle Button -->
